+++ resources/js/beacon.js
@@ -1,6 +1,12 @@
 import { createApp } from 'vue'
 import { createStore } from 'vuex'
 import Beacon from './components/Beacon.vue'
+import { parentOrigin } from './utils'
+import { createThemeApplier, resolveBootDark, followThemeMessages } from './theme'
+
+const applyTheme = createThemeApplier(window.superbotmanWidget)
+applyTheme(resolveBootDark(window.superbotmanWidget))
+followThemeMessages(applyTheme, parentOrigin())
 
 const app = createApp(Beacon)